Move the blueprint container and the selection blueprint to the same namespace. Seems onlyt separate them in the ruleset edit composer.
